fix(lint): zero eslint warnings for CI gate
CI runs eslint --max-warnings 0 across the full client source; two
unrelated pre-existing warnings were blocking merge:

1. QuickSwitcher useMemo listed 'mod' as a dep, but only formatShortcut
   (an import) reads it — the local 'mod' constant is rendered outside
   the memo. Remove it from the dep array.

2. settings-kit.tsx exported both components (Section/Field/Toolbar)
   and style constants, tripping react-refresh/only-export-components.
   Split styles into settings-kit-styles.ts; update the four importers
   (PasskeyManager, TwoFactorManager, ApiKeyManager, AccountSettingsPage)
   to import components from settings-kit and styles from -styles.
